Refactor kirby.api.ext structure
