+++ tests/tests_api/tests_api_ext/test_api_ext_topic.py
@@ -1,12 +1,10 @@
-from unittest.mock import patch, MagicMock
 import pytest
 from hypothesis import given, strategies
 
 from kafka.structs import TopicPartition
 from kafka.consumer.fetcher import ConsumerRecord
 
-from kirby.api.ext import (
-    WebClient,
+from kirby.api.ext.topic import (
     Producer,
     kirby_value_serializer,
     parse_records,
@@ -20,68 +18,6 @@ def test_creation_of_a_kirby_topic(kirby_topic_factory):
         kirby_topic.send("Hello world")
 
         assert kirby_topic.next(timeout_ms=500) == "Hello world"
-
-
-@pytest.mark.parametrize(
-    "method", ["get", "post", "put", "delete", "head", "options"]
-)
-@patch("requests.session")
-def test_web_client_calls_requests_session_methods(session_mock, method):
-    data = {"foo": True}
-
-    method_mocked = MagicMock(
-        return_value=MagicMock(
-            status_code=200, json=MagicMock(return_value=data)
-        )
-    )
-
-    session_mock.return_value = MagicMock(**{method: method_mocked})
-
-    with WebClient(
-        "external_server", "http://some.external.server"
-    ) as web_client:
-        assert getattr(web_client, method)("an_endoint") == data
-        method_mocked.assert_called_once_with(
-            "http://some.external.server/an_endoint"
-        )
-
-
-@pytest.mark.parametrize("bad_return_value", [502, 504, 500, 501])
-@pytest.mark.parametrize(
-    "method", ["get", "post", "put", "delete", "head", "options"]
-)
-@patch("requests.session")
-def test_web_client_handle_errors_of_connection(
-    session_mock, method, bad_return_value
-):
-    data = {"foo": True}
-
-    method_mocked = MagicMock(
-        side_effect=[
-            MagicMock(
-                status_code=bad_return_value, json=MagicMock(return_value={})
-            ),
-            MagicMock(status_code=200, json=MagicMock(return_value=data)),
-        ]
-    )
-
-    session_mock.return_value = MagicMock(**{method: method_mocked})
-
-    with WebClient(
-        "external_server", "http://some.external.server"
-    ) as web_client:
-        assert getattr(web_client, method)("an_endoint") == data
-        method_mocked.assert_called_with(
-            "http://some.external.server/an_endoint"
-        )
-
-
-def test_web_client_cannot_access_session_attribute():
-    with WebClient(
-        "external_server", "http://some.external.server"
-    ) as web_client:
-        with pytest.raises(AttributeError):
-            assert web_client.headers
 
 
 @strategies.composite
